requirements.txt
BeautifulSoup==3.2.1
Django==1.4
Pillow==2.0.0
Pygments==1.6
South==0.8.2
amqp==1.0.13
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.32
boto==2.11.0
celery==3.0.23
cssutils==0.9.10
dj-database-url==0.2.1
dj-static==0.0.5
django-appconf==0.6
django-celery==3.0.23
django-crispy-forms==1.2.8
django-mptt==0.5.5
django-paging==0.2.5
django-picklefield==0.3.0
django-queued-storage==0.6
django-registration==1.0
django-social-auth==0.7.25
django-sphinx==2.2.4
django-static-compiler==0.3.3
django-storages==1.1.8
django-templatetag-sugar==0.1
django-toolbelt==0.0.1
easy-thumbnails==1.3
gunicorn==0.17.4
httpagentparser==1.2.2
httplib2==0.8
kombu==2.5.14
logan==0.5.7
mock==1.0.1
nydus==0.10.6
oauth2==1.5.211
psycopg2==2.4.2
pynliner==0.5.0
python-dateutil==2.1
python-memcached==1.53
python-openid==2.2.5
pytz==2013d
raven==3.4.1
redis==2.8.0
sentry==6.2.1
setproctitle==1.1.7
simplejson==3.3.0
six==1.4.1
sorl-thumbnail==11.12
sphinxsearch==0.1
static==0.4
wsgiref==0.1.2
